Move indy/credential_def to indy/primitives/credential_definition
